Display a followable link, not a readable one
Google results point to an URL, and displays it. But when the URL is too long,
they display a "beautified version" of it: for instance,

http://superuser.com/questions/47192/google-search-from-linux-terminal
which is a long URL, is simply displayed as
superuser.com/questions/.../google-search-from-linux-terminal

While this is pretty, this isn't useful on a CLI tool: if you want to act upon
a certain result, you'll need the complete URL, not a prettified but lossy
version of it.

The search tool was showing the "display version" of the link instead of the
actual link, this patch changes that.